requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
Markdown==2.6.11
python-decouple==3.1
pytz==2018.3
